Feature/more nextflow mods (#1101)
Working on #1041
Also updated to swagger 2.3.0 while we're at it (which introduces some style changes) and some refactoring to get prepared for DAG rendering on NextFlow when we're ready

Still needed:

    DAG support
    CLI workflow execution
