+++ dockstore-client/pom.xml
@@ -20,7 +20,7 @@
     <parent>
         <groupId>io.dockstore</groupId>
         <artifactId>dockstore</artifactId>
-        <version>1.4.0-alpha.5-SNAPSHOT</version>
+        <version>1.4.0-alpha.6-SNAPSHOT</version>
     </parent>
     <artifactId>dockstore-client</artifactId>
     <packaging>jar</packaging>
@@ -42,17 +42,17 @@
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>dockstore-common</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>swagger-java-client</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>dockstore-file-plugin-parent</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.apache.avro</groupId>
