+++ dockstore-client/src/main/java/io/dockstore/client/cli/SearchClient.java
@@ -26,7 +26,7 @@ import com.google.gson.GsonBuilder;
 import com.google.gson.annotations.SerializedName;
 import io.dockstore.common.ToolWorkflowDeserializer;
 import io.swagger.client.ApiException;
-import io.swagger.client.api.ExtendedGA4GHApi;
+import io.swagger.client.api.ExtendedGa4GhApi;
 import io.swagger.client.model.ToolV1;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -61,7 +61,7 @@ public final class SearchClient {
     /**
      * @param args
      */
-    static boolean handleCommand(List<String> args, ExtendedGA4GHApi api) {
+    static boolean handleCommand(List<String> args, ExtendedGa4GhApi api) {
         String[] argv = args.toArray(new String[args.size()]);
         JCommander jc = new JCommander();
 
@@ -105,7 +105,7 @@ public final class SearchClient {
 
     }
 
-    private static boolean search(ExtendedGA4GHApi api) {
+    private static boolean search(ExtendedGa4GhApi api) {
         // this needs to be improved, obviously a hardcoded search query is not what we'll want in the future
         String body = "{\"aggs\":{\"registry_0\":{\"terms\":{\"field\":\"registry\",\"size\":5}},\"author_1\":{\"terms\":{\"field\":\"author\",\"size\":10}}},\"query\":{\"match_all\":{}}}";
         try {
@@ -126,7 +126,7 @@ public final class SearchClient {
         return true;
     }
 
-    private static boolean index(ExtendedGA4GHApi api) {
+    private static boolean index(ExtendedGa4GhApi api) {
         try {
             api.toolsIndexGet();
         } catch (ApiException e) {
