+++ dockstore-common/src/main/java/io/dockstore/common/DescriptorLanguage.java
@@ -23,7 +23,7 @@ package io.dockstore.common;
  */
 public enum DescriptorLanguage {
     // Add new descriptor language here
-    CWL("CWL", "Common Workflow Language"), WDL("WDL", "Workflow Description Language");
+    CWL("CWL", "Common Workflow Language"), WDL("WDL", "Workflow Description Language"), NEXTFLOW("NEXTFLOW", "Nextflow");
 
     /**
      * this name is used in the workflow path
