+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/BaseIT.java
@@ -15,15 +15,27 @@
  */
 package io.dockstore.client.cli;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.concurrent.TimeoutException;
+
 import io.dockstore.common.CommonTestUtilities;
 import io.dockstore.common.ConfidentialTest;
+import io.dockstore.common.Constants;
+import io.dockstore.common.Utilities;
 import io.dockstore.webservice.DockstoreWebserviceApplication;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
 import io.dropwizard.testing.DropwizardTestSupport;
+import io.swagger.client.ApiClient;
+import org.apache.commons.configuration2.INIConfiguration;
+import org.apache.commons.dbutils.handlers.ScalarHandler;
+import org.apache.commons.io.FileUtils;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
+
+import static io.dockstore.common.CommonTestUtilities.getTestingPostgres;
 
 /**
  * Base integration test class
@@ -49,5 +61,22 @@ public class BaseIT {
     @Before
     public void resetDBBetweenTests() throws Exception {
         CommonTestUtilities.cleanState(SUPPORT);
+    }
+
+    /**
+     * Shared convenience method
+     * @return
+     * @throws IOException
+     * @throws TimeoutException
+     */
+    protected static ApiClient getWebClient() {
+        final CommonTestUtilities.TestingPostgres testingPostgres = getTestingPostgres();
+        File configFile = FileUtils.getFile("src", "test", "resources", "config2");
+        INIConfiguration parseConfig = Utilities.parseConfig(configFile.getAbsolutePath());
+        ApiClient client = new ApiClient();
+        client.setBasePath(parseConfig.getString(Constants.WEBSERVICE_BASE_PATH));
+        client.addDefaultHeader("Authorization", "Bearer " + (testingPostgres
+            .runSelectStatement("select content from token where tokensource='dockstore';", new ScalarHandler<>())));
+        return client;
     }
 }
