+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/SystemClientIT.java
@@ -40,7 +40,7 @@ import io.swagger.client.ApiClient;
 import io.swagger.client.ApiException;
 import io.swagger.client.api.ContainersApi;
 import io.swagger.client.api.ContainertagsApi;
-import io.swagger.client.api.GA4GHV1Api;
+import io.swagger.client.api.Ga4Ghv1Api;
 import io.swagger.client.api.UsersApi;
 import io.swagger.client.api.WorkflowsApi;
 import io.swagger.client.auth.ApiKeyAuth;
@@ -87,6 +87,8 @@ public class SystemClientIT {
 
     public static final String QUAY_IO_TEST_ORG_TEST6 = "quay.io/test_org/test6";
     public static final String REGISTRY_HUB_DOCKER_COM_SEQWARE_SEQWARE = "registry.hub.docker.com/seqware/seqware/test5";
+    public static final DropwizardTestSupport<DockstoreWebserviceConfiguration> SUPPORT = new DropwizardTestSupport<>(
+        DockstoreWebserviceApplication.class, CommonTestUtilities.CONFIG_PATH);
 
     @Rule
     public final ExpectedSystemExit systemExit = ExpectedSystemExit.none();
@@ -96,10 +98,6 @@ public class SystemClientIT {
 
     @Rule
     public final SystemErrRule systemErrRule = new SystemErrRule().enableLog().muteForSuccessfulTests();
-
-
-    public static final DropwizardTestSupport<DockstoreWebserviceConfiguration> SUPPORT = new DropwizardTestSupport<>(
-        DockstoreWebserviceApplication.class, CommonTestUtilities.CONFIG_PATH);
 
     @BeforeClass
     public static void dumpDBAndCreateSchema() throws Exception {
@@ -191,7 +189,7 @@ public class SystemClientIT {
         assertTrue(containers.size() == 5);
 
         DockstoreTool container = containersApi.getContainerByToolPath("quay.io/test_org/test2");
-        assertFalse(container.getIsPublished());
+        assertFalse(container.isIsPublished());
 
         long containerId = container.getId();
 
@@ -291,7 +289,7 @@ public class SystemClientIT {
     @Test
     public void testGA4GHMetadata() throws IOException, TimeoutException, ApiException {
         ApiClient client = getAdminWebClient();
-        GA4GHV1Api toolApi = new GA4GHV1Api(client);
+        Ga4Ghv1Api toolApi = new Ga4Ghv1Api(client);
         final MetadataV1 metadata = toolApi.metadataGet();
         assertTrue(metadata.getFriendlyName().contains("Dockstore"));
     }
@@ -299,7 +297,7 @@ public class SystemClientIT {
     @Test
     public void testGA4GHListContainers() throws IOException, TimeoutException, ApiException {
         ApiClient client = getAdminWebClient();
-        GA4GHV1Api toolApi = new GA4GHV1Api(client);
+        Ga4Ghv1Api toolApi = new Ga4Ghv1Api(client);
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
         DockstoreTool c = getContainer();
@@ -320,7 +318,7 @@ public class SystemClientIT {
     @Test
     public void testGetSpecificTool() throws IOException, TimeoutException, ApiException {
         ApiClient client = getAdminWebClient();
-        GA4GHV1Api toolApi = new GA4GHV1Api(client);
+        Ga4Ghv1Api toolApi = new Ga4Ghv1Api(client);
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
         DockstoreTool c = getContainer();
@@ -346,7 +344,7 @@ public class SystemClientIT {
     @Test
     public void testGetVerifiedSpecificTool() throws ApiException, IOException, TimeoutException {
         ApiClient client = getAdminWebClient();
-        GA4GHV1Api toolApi = new GA4GHV1Api(client);
+        Ga4Ghv1Api toolApi = new Ga4Ghv1Api(client);
         ContainersApi containersApi = new ContainersApi(client);
         ContainertagsApi containertagsApi = new ContainertagsApi(client);
         // register one more to give us something to look at
@@ -361,7 +359,7 @@ public class SystemClientIT {
         Tag tag = tags.get(0);
 
         // verify master branch
-        assertTrue(!tag.getVerified());
+        assertTrue(!tag.isVerified());
         assertTrue(tag.getVerifiedSource() == null);
         VerifyRequest request = SwaggerUtility.createVerifyRequest(true, "test-source");
         containertagsApi.verifyToolTag(dockstoreTool.getId(), tag.getId(), request);
@@ -369,14 +367,14 @@ public class SystemClientIT {
         // check again
         tags = containertagsApi.getTagsByPath(dockstoreTool.getId());
         tag = tags.get(0);
-        assertTrue(tag.getVerified());
+        assertTrue(tag.isVerified());
         assertTrue(tag.getVerifiedSource().equals("test-source"));
     }
 
     @Test
     public void testGetFiles() throws IOException, TimeoutException, ApiException {
         ApiClient client = getAdminWebClient();
-        GA4GHV1Api toolApi = new GA4GHV1Api(client);
+        Ga4Ghv1Api toolApi = new Ga4Ghv1Api(client);
         ContainersApi containersApi = new ContainersApi(client);
         // register one more to give us something to look at
         DockstoreTool c = getContainer();
@@ -471,14 +469,14 @@ public class SystemClientIT {
         assertTrue(containers.size() == 5);
 
         DockstoreTool container = containersApi.getContainerByToolPath("quay.io/test_org/test5");
-        assertFalse(container.getIsPublished());
+        assertFalse(container.isIsPublished());
 
         long containerId = container.getId();
 
         PublishRequest pub = SwaggerUtility.createPublishRequest(true);
 
         container = containersApi.publish(containerId, pub);
-        assertTrue(container.getIsPublished());
+        assertTrue(container.isIsPublished());
 
         containers = containersApi.allPublishedContainers();
         assertTrue(containers.size() == 2);
@@ -486,7 +484,7 @@ public class SystemClientIT {
         pub = SwaggerUtility.createPublishRequest(false);
 
         container = containersApi.publish(containerId, pub);
-        assertFalse(container.getIsPublished());
+        assertFalse(container.isIsPublished());
     }
 
     @Test
@@ -667,7 +665,7 @@ public class SystemClientIT {
     }
 
     private void starring(List<Long> containerIds, ContainersApi containersApi, UsersApi usersApi)
-            throws ApiException, IOException, TimeoutException {
+            throws ApiException {
         StarRequest request = SwaggerUtility.createStarRequest(true);
         containerIds.forEach(containerId -> {
             try {
