+++ dockstore-webservice/pom.xml
@@ -27,7 +27,7 @@
     </properties>
 
     <parent>
-        <version>1.4.0-alpha.5-SNAPSHOT</version>
+        <version>1.4.0-alpha.6-SNAPSHOT</version>
         <groupId>io.dockstore</groupId>
         <artifactId>dockstore</artifactId>
         <relativePath>../pom.xml</relativePath>
@@ -37,12 +37,12 @@
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>swagger-java-quay-client</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>dockstore-common</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>io.swagger</groupId>
@@ -459,7 +459,7 @@
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>dockstore-client</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
             <exclusions>
                 <exclusion>
                     <groupId>ch.qos.logback</groupId>
@@ -476,7 +476,7 @@
         <dependency>
             <groupId>io.dockstore</groupId>
             <artifactId>dockstore-common</artifactId>
-            <version>1.4.0-alpha.5-SNAPSHOT</version>
+            <version>1.4.0-alpha.6-SNAPSHOT</version>
             <scope>test</scope>
             <type>test-jar</type>
         </dependency>
