+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/GitLabSourceCodeRepo.java
@@ -313,7 +313,7 @@ public class GitLabSourceCodeRepo extends SourceCodeRepoInterface {
             boolean validWorkflow;
             file = new SourceFile();
 
-            validWorkflow = LanguageHandlerFactory.getInterface(type, this).isValidWorkflow(content);
+            validWorkflow = LanguageHandlerFactory.getInterface(type).isValidWorkflow(content);
 
             if (validWorkflow) {
                 file.setType(type);
