+++ dockstore-webservice/src/main/java/io/dockstore/webservice/languages/WDLHandler.java
@@ -20,11 +20,14 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import com.google.common.base.Strings;
 import com.google.common.io.Files;
 import io.dockstore.client.Bridge;
 import io.dockstore.webservice.CustomWebApplicationException;
@@ -34,8 +37,13 @@ import io.dockstore.webservice.core.Tool;
 import io.dockstore.webservice.core.Version;
 import io.dockstore.webservice.core.Workflow;
 import io.dockstore.webservice.helpers.SourceCodeRepoInterface;
+import io.dockstore.webservice.jdbi.ToolDAO;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.lang3.tuple.MutablePair;
+import org.apache.commons.lang3.tuple.MutableTriple;
+import org.apache.commons.lang3.tuple.Pair;
+import org.apache.commons.lang3.tuple.Triple;
 import org.apache.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,11 +54,6 @@ import wdl4s.parser.WdlParser;
  */
 public class WDLHandler implements LanguageHandlerInterface {
     public static final Logger LOG = LoggerFactory.getLogger(WDLHandler.class);
-    private final SourceCodeRepoInterface sourceCodeRepoInterface;
-
-    WDLHandler(SourceCodeRepoInterface sourceCodeRepoInterface) {
-        this.sourceCodeRepoInterface = sourceCodeRepoInterface;
-    }
 
     @Override
     public Entry parseWorkflowContent(Entry entry, String content) {
@@ -92,7 +95,7 @@ public class WDLHandler implements LanguageHandlerInterface {
     }
 
     @Override
-    public Map<String, SourceFile> processImports(String content, Version version) {
+    public Map<String, SourceFile> processImports(String content, Version version, SourceCodeRepoInterface sourceCodeRepoInterface) {
         Map<String, SourceFile> imports = new HashMap<>();
         SourceFile.FileType fileType = SourceFile.FileType.DOCKSTORE_WDL;
         final File tempDesc;
@@ -135,6 +138,115 @@ public class WDLHandler implements LanguageHandlerInterface {
         }
 
         return imports;
+    }
+
+    /**
+     * This method will get the content for tool tab with descriptor type = WDL
+     * It will then call another method to transform the content into JSON string and return
+     * @param mainDescName the name of the main descriptor
+     * @param mainDescriptor the content of the main descriptor
+     * @param secondaryDescContent the content of the secondary descriptors in a map, looks like file paths -> content
+     * @param type tools or DAG
+     * @param dao used to retrieve information on tools
+     * @return either a list of tools or a json map
+     */
+    @Override
+    public String getContent(String mainDescName, String mainDescriptor, Map<String, String> secondaryDescContent, LanguageHandlerInterface.Type type,
+        ToolDAO dao) {
+        // Initialize general variables
+        Bridge bridge = new Bridge();
+        bridge.setSecondaryFiles((HashMap<String, String>)secondaryDescContent);
+        String callType = "call"; // This may change later (ex. tool, workflow)
+        String toolType = "tool";
+
+        // Initialize data structures for DAG
+        Map<String, List<String>> callToDependencies; // Mapping of stepId -> array of dependencies for the step
+        List<Pair<String, String>> nodePairs = new ArrayList<>();
+        Map<String, String> callToType = new HashMap<>();
+
+        // Initialize data structures for Tool table
+        Map<String, Triple<String, String, String>> nodeDockerInfo = new HashMap<>(); // map of stepId -> (run path, docker image, docker url)
+
+        // Write main descriptor to file
+        // The use of temporary files is not needed here and might cause new problems
+        File tempMainDescriptor;
+        try {
+            tempMainDescriptor = File.createTempFile("main", "descriptor", Files.createTempDir());
+            Files.write(mainDescriptor, tempMainDescriptor, StandardCharsets.UTF_8);
+        } catch (IOException e) {
+            throw new CustomWebApplicationException("could not process wdl into DAG", HttpStatus.SC_INTERNAL_SERVER_ERROR);
+        }
+
+        // Iterate over each call, grab docker containers
+        Map<String, String> callToDockerMap = bridge.getCallsToDockerMap(tempMainDescriptor);
+
+        // Get import files
+        Map<String, String> namespaceToPath = bridge.getImportMap(tempMainDescriptor);
+
+        // Create nodePairs, callToType, toolID, and toolDocker
+        for (Map.Entry<String, String> entry : callToDockerMap.entrySet()) {
+            String callId = entry.getKey();
+            String docker = entry.getValue();
+            nodePairs.add(new MutablePair<>(callId, docker));
+            if (Strings.isNullOrEmpty(docker)) {
+                callToType.put(callId, callType);
+            } else {
+                callToType.put(callId, toolType);
+            }
+            String dockerUrl = null;
+            if (!Strings.isNullOrEmpty(docker)) {
+                dockerUrl = getURLFromEntry(docker, dao);
+            }
+
+            // Determine if call is imported
+            String[] callName = callId.replaceFirst("^dockstore_", "").split("\\.");
+
+            if (callName.length > 1) {
+                nodeDockerInfo.put(callId, new MutableTriple<>(namespaceToPath.get(callName[0]), docker, dockerUrl));
+            } else {
+                nodeDockerInfo.put(callId, new MutableTriple<>(mainDescName, docker, dockerUrl));
+            }
+        }
+
+        // Iterate over each call, determine dependencies
+        callToDependencies = bridge.getCallsToDependencies(tempMainDescriptor);
+
+        // Determine start node edges
+        for (Pair<String, String> node : nodePairs) {
+            if (callToDependencies.get(node.getLeft()).size() == 0) {
+                ArrayList<String> dependencies = new ArrayList<>();
+                dependencies.add("UniqueBeginKey");
+                callToDependencies.put(node.getLeft(), dependencies);
+            }
+        }
+        nodePairs.add(new MutablePair<>("UniqueBeginKey", ""));
+
+        // Determine end node edges
+        Set<String> internalNodes = new HashSet<>(); // Nodes that are not leaf nodes
+        Set<String> leafNodes = new HashSet<>(); // Leaf nodes
+
+        for (Map.Entry<String, List<String>> entry : callToDependencies.entrySet()) {
+            List<String> dependencies = entry.getValue();
+            internalNodes.addAll(dependencies);
+            leafNodes.add(entry.getKey());
+        }
+
+        // Find leaf nodes by removing internal nodes
+        leafNodes.removeAll(internalNodes);
+
+        List<String> endDependencies = new ArrayList<>(leafNodes);
+
+        callToDependencies.put("UniqueEndKey", endDependencies);
+        nodePairs.add(new MutablePair<>("UniqueEndKey", ""));
+
+        // Create JSON for DAG/table
+        if (type == LanguageHandlerInterface.Type.DAG) {
+            return setupJSONDAG(nodePairs, callToDependencies, callToType, nodeDockerInfo);
+        } else if (type == LanguageHandlerInterface.Type.TOOLS) {
+            return getJSONTableToolContent(nodeDockerInfo);
+        }
+
+        return null;
     }
 
     /**
