+++ dockstore-webservice/src/test/java/core/CWLParseTesting.java
@@ -35,7 +35,7 @@ public class CWLParseTesting {
     @Test
     public void testOldMetadataExample() throws IOException {
         String filePath = ResourceHelpers.resourceFilePath("metadata_example0.cwl");
-        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL, null);
+        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL);
         Entry entry = sInterface.parseWorkflowContent(new Tool(), FileUtils.readFileToString(new File(filePath), StandardCharsets.UTF_8));
         Assert.assertTrue("incorrect author", entry.getAuthor().equals("Keiran Raine"));
         Assert.assertTrue("incorrect email", entry.getEmail().equals("keiranmraine@gmail.com"));
@@ -44,7 +44,7 @@ public class CWLParseTesting {
     @Test
     public void testNewMetadataExample() throws IOException {
         String filePath = ResourceHelpers.resourceFilePath("metadata_example2.cwl");
-        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL, null);
+        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL);
         Entry entry = sInterface.parseWorkflowContent(new Tool(), FileUtils.readFileToString(new File(filePath), StandardCharsets.UTF_8));
         Assert.assertTrue("incorrect author", entry.getAuthor().equals("Denis Yuen"));
         Assert.assertTrue("incorrect email", entry.getEmail().equals("dyuen@oicr.on.ca"));
@@ -53,7 +53,7 @@ public class CWLParseTesting {
     @Test
     public void testCombinedMetadataExample() throws IOException {
         String filePath = ResourceHelpers.resourceFilePath("metadata_example3.cwl");
-        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL, null);
+        LanguageHandlerInterface sInterface = LanguageHandlerFactory.getInterface(SourceFile.FileType.DOCKSTORE_CWL);
         Entry entry = sInterface.parseWorkflowContent(new Tool(), FileUtils.readFileToString(new File(filePath), StandardCharsets.UTF_8));
         Assert.assertTrue("incorrect author", entry.getAuthor().equals("Denis Yuen"));
         Assert.assertTrue("incorrect email", entry.getEmail().equals("dyuen@oicr.on.ca"));
