+++ pom.xml
@@ -19,7 +19,7 @@
 
     <groupId>io.dockstore</groupId>
     <artifactId>dockstore</artifactId>
-    <version>1.4.0-alpha.5-SNAPSHOT</version>
+    <version>1.4.0-alpha.6-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>dockstore</name>
@@ -48,6 +48,7 @@
         <maven-surefire.version>2.20</maven-surefire.version>
         <maven-failsafe.version>2.19.1</maven-failsafe.version>
         <httpcomponents.version>4.4.5</httpcomponents.version>
+        <jackson-datatype-jsr310.version>2.8.9</jackson-datatype-jsr310.version>
     </properties>
 
     <organization>
@@ -656,6 +657,11 @@
                 <artifactId>javax.ws.rs-api</artifactId>
                 <version>2.0.1</version>
             </dependency>
+            <dependency>
+                <groupId>com.fasterxml.jackson.datatype</groupId>
+                <artifactId>jackson-datatype-jsr310</artifactId>
+                <version>${jackson-datatype-jsr310.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
@@ -892,7 +898,7 @@
                 <plugin>
                     <groupId>io.swagger</groupId>
                     <artifactId>swagger-codegen-maven-plugin</artifactId>
-                    <version>2.2.3</version>
+                    <version>2.3.0</version>
                 </plugin>
                 <plugin>
                     <groupId>org.codehaus.mojo</groupId>
